Moving validation classes to private API package
