+++ platform-extras/dolphin-platform-bean-validation/src/main/resources/META-INF/mapping/validationMappingDolphinPlatform.xml
@@ -23,37 +23,37 @@
 
     <constraint-definition annotation="javax.validation.constraints.AssertFalse">
         <validated-by include-existing-validators="true">
-            <value>com.canoo.dolphin.validation.AssertFalsePropertyValidator</value>
+            <value>com.canoo.dp.impl.validation.AssertFalsePropertyValidator</value>
         </validated-by>
     </constraint-definition>
 
     <constraint-definition annotation="javax.validation.constraints.AssertTrue">
         <validated-by include-existing-validators="true">
-            <value>com.canoo.dolphin.validation.AssertTruePropertyValidator</value>
+            <value>com.canoo.dp.impl.validation.AssertTruePropertyValidator</value>
         </validated-by>
     </constraint-definition>
 
     <constraint-definition annotation="javax.validation.constraints.DecimalMax">
         <validated-by include-existing-validators="true">
-            <value>com.canoo.dolphin.validation.DecimalMaxPropertyValidatorForNumber</value>
+            <value>com.canoo.dp.impl.validation.DecimalMaxPropertyValidatorForNumber</value>
         </validated-by>
     </constraint-definition>
 
     <constraint-definition annotation="javax.validation.constraints.DecimalMin">
         <validated-by include-existing-validators="true">
-            <value>com.canoo.dolphin.validation.DecimalMinPropertyValidatorForNumber</value>
+            <value>com.canoo.dp.impl.validation.DecimalMinPropertyValidatorForNumber</value>
         </validated-by>
     </constraint-definition>
 
     <constraint-definition annotation="javax.validation.constraints.NotNull">
         <validated-by include-existing-validators="true">
-            <value>com.canoo.dolphin.validation.NotNullPropertyValidator</value>
+            <value>com.canoo.dp.impl.validation.NotNullPropertyValidator</value>
         </validated-by>
     </constraint-definition>
     
     <constraint-definition annotation="javax.validation.constraints.Null">
         <validated-by include-existing-validators="true">
-            <value>com.canoo.dolphin.validation.NullPropertyValidator</value>
+            <value>com.canoo.dp.impl.validation.NullPropertyValidator</value>
         </validated-by>
     </constraint-definition>
