+++ platform-extras/dolphin-platform-bean-validation/src/test/java/com/canoo/dp/impl/validation/ValidationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.validation;
+package com.canoo.dp.impl.validation;
 
 import org.testng.annotations.Test;
